Grocery Store Management System
A supermarket can use this tool to monitor all management and billing activities. It keeps track of all the clients, workers, and goods. Customers, staff, and items may all be added to, updated, and the specifics of numerous entries can be seen. Additionally, a client may use it to create, amend, view, and check out their shopping basket with a bill that includes all of their personal information, the date, time, and mode of payment for the purchase.